requirements.txt
captcha==0.5.0
email-validator==2.1.0.post1
Pillow==9.0.1
python-dotenv==1.0.0
streamlit==1.28.0
